requirements.txt
livekit-agents~=1.0
livekit-plugins-assemblyai~=1.0
python-dotenv
requests
